test_case_generator/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
transformers==4.37.2
python-dotenv==1.0.1
pyyaml==6.0.1
openai==1.12.0
torch==2.2.0
python-multipart==0.0.9